[refactor #14] UserDetail에서 담는 정보 수정
UserDetail에서 담는 정보 loginId -> userId
