+++ src/main/java/com/freepath/devpath/common/auth/domain/RefreshToken.java
@@ -1,4 +1,4 @@
-package com.freepath.devpath.common.auth.entity;
+package com.freepath.devpath.common.auth.domain;
 
 import lombok.*;
