+++ src/main/java/com/freepath/devpath/common/config/RedisConfig.java
@@ -1,6 +1,6 @@
 package com.freepath.devpath.common.config;
 
-import com.freepath.devpath.common.auth.entity.RefreshToken;
+import com.freepath.devpath.common.auth.domain.RefreshToken;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
