+++ src/main/java/com/freepath/devpath/common/jwt/JwtTokenProvider.java
@@ -32,11 +32,11 @@ public class JwtTokenProvider {
     }
 
     // access token 생성 메소드
-    public String createToken(String username, String role) {
+    public String createToken(String loginId, String role) {
         Date now = new Date();
         Date expiryDate = new Date(now.getTime() + jwtExpiration);
         return Jwts.builder()
-                .subject(username)
+                .subject(loginId)
                 .claim("role", role)
                 .issuedAt(now)
                 .expiration(expiryDate)
@@ -45,11 +45,11 @@ public class JwtTokenProvider {
     }
 
     // refresh token 생성 메소드
-    public String createRefreshToken(String username, String role) {
+    public String createRefreshToken(String loginId, String role) {
         Date now = new Date();
         Date expiryDate = new Date(now.getTime() + jwtRefreshExpiration);
         return Jwts.builder()
-                .subject(username)
+                .subject(loginId)
                 .claim("role", role)
                 .issuedAt(now)
                 .expiration(expiryDate)
